Perun/Adjudicator: Extract Adjudicator into dedicated module
